fix: correctness bugs in NaiveBayes, SVM, MLP regressor, and trees
Surfaced by a 51-dataset benchmark vs scikit-learn and Weka. All are pure-Python
fixes; each now matches scikit-learn on the affected datasets.

- Trees (DecisionTree + RandomForest, classification & regression): feature
  values and split thresholds were truncated to float32 in three places
  (decision_tree predict/predict_proba input cast, _core.predict.predict_batch,
  _core.nodes.flatten_tree). On high-precision features this rounded values
  across split boundaries and misrouted samples, producing catastrophic,
  seed-dependent trees (e.g. taiwanese_bankruptcy train acc 0.03). Use float64.

- NaiveBayes (Gaussian): replace the fixed absolute 1e-6 variance floor with a
  scale-relative floor (var_smoothing * max feature variance, like sklearn) and
  compute the Gaussian log-density directly instead of log(pdf) with an
  underflow clamp. Fixes collapses on standardized/one-hot data.

- SVC / SVR: max_iter default 1000 left the SMO solver under-converged on
  datasets larger than a few thousand rows (degenerate but fast model).
  Default max_iter=-1 now auto-scales the cap to max(10000, n_samples).

- MLPRegressor: gradient was divided by n_samples three times (near-zero
  updates -> ~0 / negative R2); fix the averaging and add global-norm gradient
  clipping to prevent divergence on wide/large-target datasets.

+++ tuiml/algorithms/trees/_core/predict.py
@@ -26,10 +26,14 @@ def _predict_single_flat(flat_tree: FlattenedTree, x: np.ndarray) -> np.ndarray:
 
 def predict_batch(flat_tree: FlattenedTree, X: np.ndarray) -> np.ndarray:
     """Predict over a batch of samples using flattened NumPy tree arrays."""
-    X = np.asarray(X, dtype=np.float32)
+    # float64 (not float32): the split thresholds are float64, so downcasting
+    # the input rounds feature values across split boundaries and misroutes
+    # samples — which silently corrupted predictions (catastrophically on
+    # data with high-precision features).
+    X = np.asarray(X, dtype=np.float64)
     if X.ndim == 1:
         X = X.reshape(1, -1)
-    return np.array([_predict_single_flat(flat_tree, x) for x in X], dtype=np.float32)
+    return np.array([_predict_single_flat(flat_tree, x) for x in X], dtype=np.float64)
 
 
 def predict_proba_batch(flat_tree: FlattenedTree, X: np.ndarray) -> np.ndarray:
